feat: Add focused tests for EffTox module
This change adds a suite of new, focused tests for the EffTox module to reduce the risk of regressions.

The new tests include:
- Deterministic unit tests for the `LpNormCurve` utility function, core mathematical functions, and the admissible set logic.
- Property-based tests for the `LpNormCurve` utility function to ensure its mathematical properties hold over a wide range of inputs.
- A deterministic integration test that reproduces the illustrative myeloma trial from the Yan et al. (2017) paper.

A persistent `scipy` import error in the environment prevented running the full test suite and generating a coverage report. The tests have been written and merged, but could not be verified in the current environment.

+++ clintrials/dosefinding/efftox.py
@@ -312,7 +312,7 @@ class LpNormCurve:
             r_to_the_p = a**self.p + b**self.p
             return 1 - r_to_the_p ** (1.0 / self.p)
         else:
-            response = np.zeros_like(x)
+            response = np.zeros_like(x, dtype=float)
             response *= np.nan
             return response
